requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
python-docx>=1.1.0
PyPDF2>=3.0.0
pdfplumber>=0.9.0
reportlab>=4.0.0
pyperclip>=1.8.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
streamlit-extras>=0.3.0